Day 02
No need to brute force it. The system only has addition and
multiplication in it so the output is pretty determined. Just find the
coarse scale and the fine scale (which turned out to be 1) and then
find the combinations that add up to the goal.

Price is right rules for the coarse and then make up the difference
with the other. Would have been perhaps more difficult if they were
relatively prime or the fine adjustment was a bit larger than one.